last commit for monday
my code broke. I was painstaikingly checking if every single piece of css i wrote was breaking the code by changing backgr color to see if it updated. In that way i could go back and remove the code. but now i wrote to many changes in one go to know WHAT broke my code. so i give up for today and I really don't think i can find the culprit as my bugconsole says all is good. So no more driving blind for the index. it is what it is now and I'm handing it in as is...